Slight refactor here.
Added new type to service generator.

It's clear that cloud variables don't really fit into the translations model, will have to think about how to work this in.

The share service will not be used in favour of a lower level interface. Namespaces are hashed as well as the variable names to reduce packet size, and provide some means of separation between variables. It will become clear how this fits into a rest model tomorrow, probably ;)